test/daily-tasks/requirements.txt
cffi==1.15.1
cryptography==38.0.1
pycparser==2.21
PyYAML==6.0
